fix normalize markdown
